+++ tubular/aggregations.py
@@ -1,7 +1,7 @@
 """Contains transformers for performing data aggregations."""
 
 from enum import Enum
-from typing import Union
+from typing import Any, Union
 
 import narwhals as nw
 from beartype import beartype
@@ -11,6 +11,7 @@ from beartype.vale import Is
 from tubular._utils import (
     _convert_dataframe_to_narwhals,
     _return_narwhals_or_native_dataframe,
+    block_from_json,
 )
 from tubular.base import BaseTransformer, register
 from tubular.mixins import DropOriginalMixin
@@ -117,7 +118,7 @@ class BaseAggregationTransformer(BaseTransformer, DropOriginalMixin):
 
     FITS = False
 
-    jsonable = False
+    jsonable = True
 
     @beartype
     def __init__(
@@ -150,6 +151,44 @@ class BaseAggregationTransformer(BaseTransformer, DropOriginalMixin):
         self.aggregations = aggregations
 
         self.drop_original = drop_original
+
+    @block_from_json
+    def to_json(self) -> dict[str, Any]:
+        """Dump transformer to json dict.
+
+        Returns:
+        -------
+        dict[str, Any]:
+            jsonified transformer. Nested dict containing levels for attributes
+            set at init and fit.
+
+
+        Example:
+        -------
+        ```pycon
+        >>> baseAggregationTransformer = BaseAggregationTransformer(
+        ...     columns="a",
+        ...     aggregations=["min", "max"],
+        ... )
+        >>> baseAggregationTransformer.to_json()  # doctest: +NORMALIZE_WHITESPACE
+        {'tubular_version': ...,
+        'classname': 'BaseAggregationTransformer',
+        'init': {'columns': ['a'],
+        'copy': False,
+        'verbose': False,
+        'return_native': True,
+        'aggregations': ['min', 'max'],
+        'drop_original': False},
+        'fit': {}}
+
+        ```
+
+        """
+        json_dict = super().to_json()
+        json_dict["init"].update(
+            {"aggregations": self.aggregations, "drop_original": self.drop_original}
+        )
+        return json_dict
 
     @beartype
     def transform(
@@ -479,7 +518,7 @@ class AggregateColumnsOverRowTransformer(BaseAggregationTransformer):
 
     FITS = False
 
-    jsonable = False
+    jsonable = True
 
     @beartype
     def __init__(